Move Patterns.kt into utils package of core module
